Alright here's the first real commit...
Basically this adds the Rails 4.0.2 project, all of the backbone.js / underscore.js / marionette.js assets.

Creates a basic model Contact which stores "contact information". This is the basic item that this tiny project CRUDs.

Sets up a RESTful resource route in routes.rb for the Contacts model. Sets up a contacts controller to handle the JSON REST API.

Sets up the single page that this app operates on: routes that page to "/".

Defines the client-side Model and Collection classes for the contacts using Backbone.js.

Creates a marionette.js Application object, used to init/fetch all the necessary data, and then render it when ready.

Uses the marionette CollectionView and ItemView classes in conjunction to render the Contacts list.

...A fine start! Hopefully will be able to finish the rest of the CRUD functionality by the end of today.