Adds probably 90% of the difficult CSS styling dealing with spacing and divs and fonts everything everything looking squeaky clean and minimal and elegant and 2014. #winning
Sort of creates the look of a collection of virtual "contact cards".

Uses a lot of divs and floats (as is the norm); adds some new ones where further separation is necessary.

Adds a whole lot to the contact_list.css file, as well as comments.

Right aligns the text inputs and the text-display cells, with the exception of the "notes" area.

Adds small-caps arial fonts globally.

There's really only a very small amount left to do on this project before it's over.