Organizes things better, puts the name in the top left, followed by the picture. Cleans up a lot of the excess in the interface. Makes it so the "Add contact" popup shares the same css layout as the list display.
Generally tweaks the css and markup of many-many elements to make things look generally cleaner and/or more usable.

Centers the page header and the Contact list.

Puts the delete link up in the upper-right corner of each list-item, as a windows-style red X.

Adds a fade-out effect on deletion.

Also this makes it so the "add contact" window closes after a contact is added.

Overall, thingslook pretty good so far. I'm almost ready to submit this. Remaining tasks include:

- Change the State text input for a select box with the 50 states.
- Add a length validation to the zip-code inputs
- Add an email validation
- Make it so that any of these values can be editted in-place.
- Add a phone number field / validation.