In this commit I incorporate the jquery.jeditable plugin such that all of the values are editable on clicking, instead of via an "edit mode" / "display mode" sort of setup.. Uses the existing RESTful Backbone.js client->server sync to update/add the necessary data on the server.
This applies both to editing an existing Contact or adding a new one, which really gives adding a contact the UI "feel" of filling out a physical contact card. This really completes the "cleaning-up" of the user interface.

Tweaks a lot more CSS stuff to give this just the right look / fixes some alignment problems...

Tweaks some internal stuff in the jquery.jeditable source file itself: firstly to make it so that an edit-input can be "blur'd" by opening another jquery.jeditable element. This is important to facilitate being able to quickly fill in these inputs by clicking through them, one after another..

Also, changed some things so that the default text when nothing is saved in the model for a particular attribute is available for CSS styling. Very important to give that default text a distinct visual look.

Finally, makes the "State" input a select drop-down input preloaded with the 50 U.S. states (instead of just a simple text input). This too fits into the jquery.jeditable plugin system.