Adds an "Active State" to the contact list, making it so that a contact card is visually "selected" when a user is interacting with it.
Renames the SingleContactView class "ContactCardView", to more accurately communicate the purpose of that class.

Adds a setActive/Inactive function to the ContactCardView class. The Active state is triggered by the mousedown event on the View's $el. The inactive state set by a mousedown event binded to the topmost HTML object.

Transitions of the active state are accompanied by a nice JQuery animation which involves a box shadow. Adds the jquery.animate-shadow plugin to accomodate the shadow animation behavior.

Despite the lack of validations and other small imperfections (the inability to use the <TAB> key to loop through the form inputs), this thing is pretty much done...