Rearranges some of the code to make it cleaner.
Basically just overrides the ContactCardView render method, such that it will call the setupJQueryPlugins() method after each render. Previously this would manually be called from a variety of callbacks and "reRender" functions, and so ultimately this is much cleaner in that it makes the code much simpler and more readable.

Cleans up all of those callbacks and connectors and reRender functions, all of which just served the purpose of setting up the 3rd party JQuery plugins after template rendering... Now this happens as a normal part of the Marionette.ItemView rendering pipeline.

Had to do some fancy footwork to manually render the template within the 'render' function, and to correctly pass the model data and the templateHelpers to the ContactCard template.

Also, moves the listAggregate handler definition into the Marionette.Application onStart callback.

Adds comments to the server-side code, like routes.rb and contacts_controller.rb